Simplify the conditional compilation for the JPEG 2000 code stream format decoder
